# Restrict intermediate queries to population