--- a/tests/integration/measures/test_calculate.py
+++ b/tests/integration/measures/test_calculate.py
@@ -10,6 +10,11 @@ from ehrql.tables import EventFrame, PatientFrame, Series, table
 @table
 class patients(PatientFrame):
     sex = Series(str)
+
+
+@table
+class addresses(EventFrame):
+    date = Series(date)
     region = Series(str)
 
 
@@ -26,6 +31,7 @@ def test_get_measure_results(engine):
     foo_event_count = events_in_interval.where(events.code == "foo").count_for_patient()
     had_event = events_in_interval.exists_for_patient()
     event_value = events_in_interval.value.sum_for_patient()
+    region = addresses.sort_by(addresses.date).last_for_patient().region
 
     intervals = years(3).starting_on("2020-01-01")
     measures = Measures()
@@ -41,7 +47,7 @@ def test_get_measure_results(engine):
         "foo_events_by_region",
         numerator=foo_event_count,
         denominator=event_count,
-        group_by=dict(region=patients.region),
+        group_by=dict(region=region),
         intervals=intervals,
     )
     measures.define_measure(
@@ -55,7 +61,7 @@ def test_get_measure_results(engine):
         "event_value_by_region",
         numerator=event_value,
         denominator=patients.exists_for_patient(),
-        group_by=dict(region=patients.region),
+        group_by=dict(region=region),
         intervals=intervals,
     )
     measures.define_measure(
@@ -64,7 +70,7 @@ def test_get_measure_results(engine):
         denominator=patients.exists_for_patient(),
         group_by=dict(
             sex=patients.sex,
-            region=patients.region,
+            region=region,
         ),
         intervals=intervals,
     )
@@ -75,13 +81,17 @@ def test_get_measure_results(engine):
         intervals=intervals,
     )
 
-    patient_data, event_data = generate_data(intervals)
-    engine.populate({patients: patient_data, events: event_data})
+    patient_data, address_data, event_data = generate_data(intervals)
+    engine.populate(
+        {patients: patient_data, addresses: address_data, events: event_data}
+    )
 
     results = get_measure_results(engine.query_engine(), measures)
     results = list(results)
 
-    expected = calculate_measure_results(intervals, patient_data, event_data)
+    expected = calculate_measure_results(
+        intervals, patient_data, address_data, event_data
+    )
     expected = list(expected)
 
     # We don't care about the order of the results
@@ -95,10 +105,21 @@ def generate_data(intervals):
         dict(
             patient_id=patient_id,
             sex=rnd.choice(["male", "female"]),
-            region=rnd.choice(["London", "The North", "The Countryside"]),
         )
         for patient_id in range(1, 50)
     ]
+    # Generate some addresses (at least one) for each patient
+    address_data = []
+    interval_range = (intervals[0][0], intervals[-1][1])
+    for patient in patient_data:
+        for _ in range(rnd.randint(1, 3)):
+            address_data.append(
+                dict(
+                    patient_id=patient["patient_id"],
+                    date=random_date_in_interval(rnd, interval_range),
+                    region=rnd.choice(["London", "The North", "The Countryside"]),
+                )
+            )
     # For each interval and patient, generate some events (possibly zero)
     event_data = []
     for interval in intervals:
@@ -114,7 +135,7 @@ def generate_data(intervals):
                 )
                 for _ in range(event_count)
             )
-    return patient_data, event_data
+    return patient_data, address_data, event_data
 
 
 def random_date_in_interval(rnd, interval):
@@ -123,11 +144,13 @@ def random_date_in_interval(rnd, interval):
     return interval[0] + timedelta(days=offset)
 
 
-def calculate_measure_results(intervals, patient_data, event_data):
+def calculate_measure_results(intervals, patient_data, address_data, event_data):
     nums = defaultdict(int)
     dens = defaultdict(int)
 
-    for interval, patient, events in group_events(intervals, patient_data, event_data):
+    for interval, patient, address, events in group_events(
+        intervals, patient_data, address_data, event_data
+    ):
         event_count = len(events)
         foo_count = len([e for e in events if e["code"] == "foo"])
         had_event = 1 if events else 0
@@ -135,19 +158,19 @@ def calculate_measure_results(intervals, patient_data, event_data):
 
         nums[("foo_events_by_sex", interval, patient["sex"], None)] += foo_count
         dens[("foo_events_by_sex", interval, patient["sex"], None)] += event_count
-        nums[("foo_events_by_region", interval, None, patient["region"])] += foo_count
-        dens[("foo_events_by_region", interval, None, patient["region"])] += event_count
+        nums[("foo_events_by_region", interval, None, address["region"])] += foo_count
+        dens[("foo_events_by_region", interval, None, address["region"])] += event_count
         nums[("had_event_by_sex", interval, patient["sex"], None)] += had_event
         dens[("had_event_by_sex", interval, patient["sex"], None)] += 1
         nums[
-            ("event_value_by_region", interval, None, patient["region"])
+            ("event_value_by_region", interval, None, address["region"])
         ] += event_value
-        dens[("event_value_by_region", interval, None, patient["region"])] += 1
+        dens[("event_value_by_region", interval, None, address["region"])] += 1
         nums[
-            ("had_event_by_sex_and_region", interval, patient["sex"], patient["region"])
+            ("had_event_by_sex_and_region", interval, patient["sex"], address["region"])
         ] += had_event
         dens[
-            ("had_event_by_sex_and_region", interval, patient["sex"], patient["region"])
+            ("had_event_by_sex_and_region", interval, patient["sex"], address["region"])
         ] += 1
         nums[("foo_events", interval, None, None)] += foo_count
         dens[("foo_events", interval, None, None)] += event_count
@@ -168,14 +191,19 @@ def calculate_measure_results(intervals, patient_data, event_data):
         )
 
 
-def group_events(intervals, patient_data, event_data):
+def group_events(intervals, patient_data, address_data, event_data):
     "Group events by interval and patient"
     for patient in patient_data:
         patient_events = [
             e for e in event_data if e["patient_id"] == patient["patient_id"]
         ]
+        patient_addresses = sorted(
+            [a for a in address_data if a["patient_id"] == patient["patient_id"]],
+            key=lambda a: a["date"],
+        )
+        address = patient_addresses[-1]
         for interval in intervals:
             interval_events = [
                 e for e in patient_events if interval[0] <= e["date"] <= interval[1]
             ]
-            yield interval, patient, interval_events
+            yield interval, patient, address, interval_events