Previously we ran every intermediate query against all the patients that happened to be in the table being queried and then only filtered the results down to the defined population when combining the results at the end.

This has the advantage of simplicity but it does mean the intermediate results tables are larger than they need to be and, potentially, the intermediate queries are slower. It also means that the database has to evaluate the population inclusion logic at the same time as doing a very wide join across all the intermediate tables which creates a big burst of CPU pressure right at the final query.

This PR changes the logic so we first build a table containing just the patient IDs we want in our population and then use this to restrict the matching patients in any subsequent queries. The final wide join thus requires no WHERE clause.

Where the query population is a small fraction of the total population in the database this can result in some dramatic speedups. As the fraction of the population increases this improvement diminishes and where the population includes all (or almost all) possible patients including this condition will make things slower. Initial testing seems to show that the speedups are so significant, and the slowdowns rare and mild enough, that this is still a net benefit but we'll need to keep this under review.

For interest, here is the original year-old Slack thread in which this approach was discussed:
https://bennettoxford.slack.com/archives/C03FB777L1M/p1657899866106729?thread_ts=1657897715.091099&cid=C03FB777L1M