@evansd 2023-08-14
Here are some initial performance comparisons:

Job | Population | Original Runtime | New Runtime
-- | -- | -- | --
[OpenPrompt 1](https://jobs.opensafely.org/the-healthcare-utilisation-of-people-with-long-covid-openprompt/healthcare_utilisation_openprompt/19635/k3tlmdgtryog65j7/) | 1,200 | 2 hours | 8 minutes
[OpenPrompt 2](https://jobs.opensafely.org/the-healthcare-utilisation-of-people-with-long-covid-openprompt/healthcare_utilisation_openprompt/19687/lrtplflb2dx3mick/) | 6,900 | 44 hours | 36 minutes
[Diabetes](https://jobs.opensafely.org/impact-of-covid-19-on-diabetes-care-in-england/covid-diabetes-care/19560/sbweasdv44dbmtt6/) | 26 million | 44 minutes | 57 minutes

As you can see, for jobs with small populations there is a very significant speed up.

For jobs with large populations – in this case `(age < 110) & current_registration.exists_for_patient()` – there is some slowdown, but bear in mind these are wall-clock times and subject to noise according to what else the database is having to do at the time. We should [improve our logging of query stats](https://github.com/opensafely-core/ehrql/issues/1444) so that we can make some direct comparisons of CPU cost.

It would also be good to parse population sizes out of our historical job logs so we can determine the distribution of population sizes which would help us estimate the impact (whether positive or negative) of this change.

A further option here is to be selective about when we enable this behaviour so that we only use it where we expect a positive impact. There are a few possibilities:

 * **Build the population table; count its rows; use that count to determine the best behaviour**
   This is the most correct approach, but adds significant architectural complexity. We'd no longer have separate compile and execute phases and, as determining the thresholds would be backend-specific, the query engine would need to involve the backend somehow in this logic.

 * **Use some heuristic for identifying queries we expect to benefit**
   The one that springs immediately to mind is "does the population definition restrict the population to an inline table?". This should be relatively easy to determine, and it's hard to see circumstances under which this obtains and the new query approach isn't a massive benefit. Of course, there may be other kinds of query which would also benefit and we'll miss those. But maybe that's OK for now? If we come across other queries which would benefit we can try to identify the patterns they identify and extent our heuristic accordingly.

 * **Allow the user to explicitly opt in to this behaviour somehow**
   This is in some ways the simplest, but I worry deeply about going down this road.