@evansd 2023-08-15
I suppose I should have qualified that as "use some _purely syntactic_ heuristic" i.e. one that doesn't depend on what any of the tables actually contain. Both DBAs and query planners use table stats to guide them which this approach wouldn't be able to do. But it may well be good enough for our purposes in any case.