@sebbacon 2023-08-15
Yes, that's how I understood it. I could have been clearer that I was commenting with respect to this:

> there may be other kinds of query which would also benefit and we'll miss those. But maybe that's OK for now? If we come across other queries which would benefit we can try to identify the patterns they identify and extent our heuristic accordingly.

I mean, that's all fine and normal, right?!