@evansd 2023-08-15
Ah yes, got you – and yes, it's definitely normal but right now I don't feel like the word "fine" applies to anything to with databases :)