@Jongmassey 2023-08-15
I'm not sure you've got enough actual CPU (not wall) time data to develop such a heuristic