@evansd 2023-08-15
Well my first candidate was "if the population is restricted to a inline patient table then use the new approach". I think the improvements in that case are _so_ far above the noise floor that we don't need to worry about doing anything more accurate there.

But yes, if we want to start investigating other opportunities we're definitely going to need more and better data.

Of course, we don't even have enough data yet to know if we need to be selective about applying this approach. It could be, for all we know, that it's fine to use it all the time and it doesn't appreciably slow down even queries with large populations.