@evansd 2023-08-18
Now we've got CPU time logging I was able to do a head to head comparison of the above `covid-diabetes-care` job using both the code on main and the code on this branch.

In wall-clock terms, the code on this branch was faster: 1hr 7m vs 1hr 40m.

In terms of aggregate CPU time, the code on this branch was _also_ faster: 10070.121s vs 11735.910s.

So, tentatively, I think that means all this discussion of when best to apply to approach is moot because we should just apply it all the time.

I'd like to do a more detailed comparison of the logs though to see how the performance of individual queries compares.