@evansd 2023-08-23
I've managed to find some jobs with unusually large populations where this approach is slightly slower than previously.

In both cases, the population filter doesn't include any kind of practice registration requirement: one is defined as `(age <= 100) & (age >= 18) & sex.contains("male")` and the other simply as `age > 0`. This is quite unusual in my experience as generally registration history is key for knowing over what period a patient's records can be relied upon.

The comparative stats are:

Dataset | Pop. Size | Tables | Total time (before) | Total time (after) | Query time (before) | Query time (after) | CPU time (before) | CPU time (after)
-- | -- | -- | -- | -- | -- | -- | -- | --
[shielding](https://github.com/opensafely/shielding-evaluation/blob/c6e099f8ac2f4b0788f0343a72d6e5289e18deec/analysis/datasets.py) | 27 million | 55 | 1h 7m | 1hr 9m | 35m 55s | 39m 51s | 23,498s | 29,676s
[openprompt](https://github.com/opensafely/openprompt-hrqol/blob/a7499e62ee8096745048f723b58662caa794988e/analysis/define_openprompt_baseline.py) | 34 million | 10 | 15m | 16m | 1m 58s | 4m 21s | 708s | 637s

The difference between "total time" and "query time" is the time spent downloading results which in both cases dominates the total runtime meaning there's little overall difference from the user's perspective.

Wall-clock query times are slower in both cases under the new approach, but the CPU time picture is mixed: slightly faster in one case, significantly slower in the other.

On this basis, I'm minded to say that we should merge this PR as is and not try to get clever about when we do and don't apply it. There will be some cases like the above where we underperform, but I think these will be rare, the underperformance will be relatively mild, and I've got no confidence we could accurately identify them in advance.